+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * CountIcon renamed to NotificationIcon
 * StatusBarComponent renamed to StatusBar
+* HeaderBar left and right components no longer render containers when passed custom components
 
 ### Added
 
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * CheckBox - androidRipple set of props
 * CheckBox - customIcon prop
 * DeleteButton - customIcon, showShadow and androidRipple set of props
+* HeaderBar - androidRipple set of props
 
 ### Removed